Refactor: Move swig wrappers out of __init__.py
Closes #1215.

Will also make it easier to reuse those wrappers for #1119
