+++ python/sdist/amici/swig_wrappers.py
@@ -1,0 +1,1 @@
+../../amici/swig_wrappers.py
